refactor: validator 인터페이스 제거 및 네이밍 수정
