+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -20,21 +20,21 @@ import java.util.stream.Collectors;
 public class OrderService {
     private final MenuRepository menuRepository;
     private final OrderRepository orderRepository;
-    private final OrderValidator orderValidator;
+    private final OrderTableValidator orderTableValidator;
 
     public OrderService(
             final MenuRepository menuRepository,
             final OrderRepository orderRepository,
-            final OrderValidatorImpl orderValidator
+            final OrderTableValidator orderTableValidator
     ) {
         this.menuRepository = menuRepository;
         this.orderRepository = orderRepository;
-        this.orderValidator = orderValidator;
+        this.orderTableValidator = orderTableValidator;
     }
 
     public Long order(final OrderCreateRequest request) {
         final List<Menu> menus = menuRepository.findByIdIn(request.extractMenuIds());
-        orderValidator.validate(request, menus);
+        orderTableValidator.validate(request, menus);
         final List<OrderLineItem> orderLineItems = createOrderLineItems(menus, request);
         final Order order = Order.ofCooking(request.getOrderTableId(), orderLineItems);
         final Order saveOrder = orderRepository.save(order);
